fix cross platform build error.
